Library seacoral-basics

SeaCoral is an orchestration tool that is able to combine the power of a wide range of test generation techniques in order to automatically produce tests for your C projects.

This library exposes the following toplevel modules:

Basics Miscellaneous helpers.

Format_shims.cppo

Memo

NEL