Library seacoral-cil-shims

SeaCoral is an orchestration tool that is able to combine the power of a wide range of test generation techniques in order to automatically produce tests for your C projects.

This library exposes the following toplevel modules:

Cil

Cil_pretty

Cilint